+++ src/@providers/docker/provider.ts
@@ -4,7 +4,6 @@ import { DockerCredentials, DockerCredentialsSchema } from './credentials.ts';
 import { DockerDatabaseService } from './services/database.ts';
 import { DockerDeploymentService } from './services/deployment.ts';
 import { DockerNamespaceService } from './services/namespace.ts';
-import { DockerPodService } from './services/pod.ts';
 import { DockerTaskService } from './services/task.ts';
 import { DockerVolumeService } from './services/volume.ts';
 import { DockerInfo } from './types.ts';
@@ -20,7 +19,6 @@ export default class DockerProvider extends Provider<DockerCredentials> {
     task: new DockerTaskService(this.name, this.credentials, this.providerStore),
     volume: new DockerVolumeService(this.name, this.credentials, this.providerStore),
     database: new DockerDatabaseService(this.name, this.credentials, this.providerStore),
-    pod: new DockerPodService(this.name, this.credentials, this.providerStore),
   };
 
   public async testCredentials(): Promise<boolean> {
