+++ src/@providers/kubernetes/provider.ts
@@ -5,7 +5,6 @@ import { KubernetesDeploymentService } from './services/deployment.ts';
 import { KubernetesHelmChartService } from './services/helm-chart.ts';
 import { KubernetesIngressRuleService } from './services/ingress-rule.ts';
 import { KubernetesNamespaceService } from './services/namespace.ts';
-import { KubernetesPodService } from './services/pod.ts';
 import { KubernetesServiceService } from './services/service.ts';
 
 export default class KubernetesProvider extends Provider<KubernetesCredentials> {
@@ -19,7 +18,6 @@ export default class KubernetesProvider extends Provider<KubernetesCredentials> 
     service: new KubernetesServiceService(this.name, this.credentials, this.providerStore),
     ingressRule: new KubernetesIngressRuleService(this.name, this.credentials, this.providerStore),
     helmChart: new KubernetesHelmChartService(this.name, this.credentials, this.providerStore),
-    pod: new KubernetesPodService(this.name, this.credentials, this.providerStore),
   };
 
   public async testCredentials(): Promise<boolean> {
