+++ src/@resources/input.schema.json
@@ -1288,32 +1288,6 @@
           "if": {
             "properties": {
               "type": {
-                "const": "pod",
-                "type": "string"
-              }
-            }
-          },
-          "then": {
-            "additionalProperties": false,
-            "properties": {
-              "account": {
-                "type": "string"
-              },
-              "type": {
-                "const": "pod",
-                "type": "string"
-              }
-            },
-            "required": [
-              "type"
-            ],
-            "type": "object"
-          }
-        },
-        {
-          "if": {
-            "properties": {
-              "type": {
                 "properties": {},
                 "type": "object"
               }
@@ -1713,7 +1687,6 @@
             "kubernetesCluster",
             "loadBalancer",
             "namespace",
-            "pod",
             "secret",
             "service",
             "task",
