--- src/@resources/pod/inputs.schema.json
@@ -1,10 +1,0 @@
-{
-  "$ref": "#/definitions/PodInputs",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "PodInputs": {
-      "additionalProperties": false,
-      "type": "object"
-    }
-  }
-}
