--- src/@resources/pod/inputs.ts
@@ -1,5 +1,0 @@
-export type PodInputs = {
-  labels?: Record<string, string>;
-};
-
-export default PodInputs;
