--- src/@resources/pod/outputs.ts
@@ -1,5 +1,0 @@
-export type PodOutputs = {
-  labels?: Record<string, string>;
-};
-
-export default PodOutputs;
